requirements.txt
opencv-python~=4.5.1.48
numpy~=1.19.5
matplotlib~=3.3.3
Pillow~=8.1.0
pandas~=1.2.0